fix: update stale bsgs_secondcheck/thirdcheck forward declarations in keyhunt.cpp
The forward declarations at keyhunt.cpp:565-566 still used uint32_t for
parameter 'a' after the definitions and search_context.h were updated to
uint64_t. This mismatch is undefined behavior in C++.

+++ src/keyhunt.cpp
@@ -562,8 +562,8 @@ void sleep_ms(int milliseconds);
 /* bsgs_sort, bsgs_myheapsort, bsgs_insertionsort, bsgs_introsort, bsgs_swap,
    bsgs_heapify, bsgs_partition, and bsgs_searchbinary now declared in bsgs/bsgs_sort.h */
 
-int bsgs_secondcheck(Int *start_range,uint32_t a,uint32_t k_index,Int *privatekey);
-int bsgs_thirdcheck(Int *start_range,uint32_t a,uint32_t k_index,Int *privatekey);
+int bsgs_secondcheck(Int *start_range,uint64_t a,uint32_t k_index,Int *privatekey);
+int bsgs_thirdcheck(Int *start_range,uint64_t a,uint32_t k_index,Int *privatekey);
 
 /* Vanity functions (defined in search/search_vanity.cpp) */
 bool vanityrmdmatch(unsigned char *rmdhash);
